More progress on side menu, node separation
